Project_Starter_Files-Building_a_Metrics_Dashboard/reference-app/frontend/requirements.txt
Flask==1.1.2
requests==2.23.0
pylint==2.4.0
gunicorn==20.0.4
pymongo==3.11.0
Flask-PyMongo==2.3.0
jaeger-client==4.3.0
prometheus-client==0.12.0
opentelemetry-api==0.12b0
opentelemetry-exporter-jaeger==0.12b0
opentelemetry-instrumentation-flask==0.12b0
opentelemetry-instrumentation-requests==0.12b0
opentelemetry-instrumentation==0.12b0
opentelemetry-sdk==0.12b0
prometheus-client==0.12.0
prometheus-flask-exporter==0.18.1
Flask-OpenTracing==1.1.0